Moved NetworkConnection to ConnectionHandle, NetworkConnection Initialization logic to a dedicated factory class.
