+++ src/include/network/connection_handle.h
@@ -12,57 +12,195 @@
 
 #pragma once
 #include <unordered_map>
+#include <event2/buffer.h>
+#include <event2/bufferevent.h>
+#include <event2/event.h>
+#include <event2/listener.h>
 
-#include "network_state.h"
+#include <csignal>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+#include <vector>
+
+#include <arpa/inet.h>
+#include <netinet/tcp.h>
+#include <sys/file.h>
+
+#include "common/exception.h"
 #include "common/logger.h"
+#include "type/types.h"
+
+#include "network/connection_handler_task.h"
+#include "protocol_handler.h"
+#include "marshal.h"
+#include "network_state.h"
+
+#include <openssl/ssl.h>
+#include <openssl/err.h>
+
 
 namespace peloton{
 namespace network {
 
-// TODO(tianyu): remove this front declaration once we breakup NetworkConnection
-class NetworkConnection;
-
-class ConnectionHandleStateMachine {
+class ConnectionHandle {
 public:
-  /**
-   * A state machine is defined to be a set of states, a set of symbols it supports, and a function mapping each
-   * state and symbol pair to the state it should transition to. i.e. transition_graph = state * symbol -> state
-   *
-   * In addition to the transition system, our network state machine also needs to perform actions. Actions are
-   * defined as functions (lambdas, or closures, in various other languages) and is promised to be invoked by the
-   * state machine after each transition if registered in the transition graph.
-   *
-   * So the transition graph overall has type transition_graph = state * symbol -> state * action
-   */
-  using action = Transition (*)(NetworkConnection &);
-  using transition_result = std::pair<ConnState, action>;
 
-  explicit ConnectionHandleStateMachine(ConnState state): current_state_(state) {}
+  ConnectionHandle(int sock_fd, ConnectionHandlerTask *handler);
 
   /**
-   * Runs the internal state machine, starting from the symbol given, until no more
-   * symbols are available.
-   *
-   * Each state of the state machine defines a map from a transition symbol to an action
-   * and the next state it should go to. The actions can either generate the next symbol,
-   * which means the state machine will continue to run on the generated symbol, or signal
-   * that there is no more symbols that can be generated, at which point the state machine
-   * will stop running and return, waiting for an external event (user interaction, or system event)
-   * to generate the next symbol.
-   *
-   * @param action starting symbol
-   * @param connection the network connection object to apply actions to
+   * refill_read_buffer - Used to repopulate read buffer with a fresh
+   * batch of data from the socket
    */
-  void Accept(Transition action, NetworkConnection &connection);
+  Transition FillReadBuffer();
+
+  /**
+   * Update the existing event to listen to the passed flags
+   */
+  bool UpdateEvent(short flags);
+
+  WriteState WritePackets();
+
+  std::string WriteBufferToString();
+
+  Transition CloseSocket();
+
+  void Reset();
+
+  inline void TriggerStateMachine() { state_machine_.Accept(Transition::WAKEUP, *this); }
+
+  // Exposed for testing
+  const std::unique_ptr<ProtocolHandler> &GetProtocolHandler() const {
+    return protocol_handler_;
+  }
+
+  // State Machine actions
+  Transition Wait();
+  Transition Process();
+  Transition ProcessWrite();
+  Transition GetResult();
 
 private:
+  class StateMachine {
+  public:
+    /**
+     * A state machine is defined to be a set of states, a set of symbols it supports, and a function mapping each
+     * state and symbol pair to the state it should transition to. i.e. transition_graph = state * symbol -> state
+     *
+     * In addition to the transition system, our network state machine also needs to perform actions. Actions are
+     * defined as functions (lambdas, or closures, in various other languages) and is promised to be invoked by the
+     * state machine after each transition if registered in the transition graph.
+     *
+     * So the transition graph overall has type transition_graph = state * symbol -> state * action
+     */
+    using action = Transition (*)(ConnectionHandle &);
+    using transition_result = std::pair<ConnState, action>;
+
+
+    /**
+     * Runs the internal state machine, starting from the symbol given, until no more
+     * symbols are available.
+     *
+     * Each state of the state machine defines a map from a transition symbol to an action
+     * and the next state it should go to. The actions can either generate the next symbol,
+     * which means the state machine will continue to run on the generated symbol, or signal
+     * that there is no more symbols that can be generated, at which point the state machine
+     * will stop running and return, waiting for an external event (user interaction, or system event)
+     * to generate the next symbol.
+     *
+     * @param action starting symbol
+     * @param connection the network connection object to apply actions to
+     */
+    void Accept(Transition action, ConnectionHandle &connection);
+
+  private:
+    /**
+     * delta is the transition function that defines, for each state, its behavior and the
+     * next state it should go to.
+     */
+    static transition_result Delta_(ConnState state, Transition transition);
+    ConnState current_state_ = ConnState::READ;
+  };
+
+  friend class StateMachine;
+  friend class ConnectionHandleFactory;
+
+  ProcessResult ProcessInitial();
+
   /**
-   * delta is the transition function that defines, for each state, its behavior and the
-   * next state it should go to.
+   * Extracts the header of a Postgres start up packet from the read socket buffer
    */
-  static transition_result Delta_(ConnState state, Transition transition);
-  ConnState current_state_;
+  static bool ReadStartupPacketHeader(Buffer &rbuf, InputPacket &rpkt);
+
+
+  /**
+   * Routine to deal with the first packet from the client
+   */
+  bool ProcessInitialPacket(InputPacket* pkt);
+
+  /**
+   * Routine to deal with general Startup message
+   */
+  bool ProcessStartupPacket(InputPacket* pkt, int32_t proto_version);
+
+  /**
+   * Routine to deal with SSL request message
+   */
+  bool ProcessSSLRequestPacket(InputPacket *pkt);
+
+  /**
+   * Writes a packet's header (type, size) into the write buffer
+   */
+  WriteState BufferWriteBytesHeader(OutputPacket *pkt);
+
+  /**
+   * Writes a packet's content into the write buffer
+   */
+  WriteState BufferWriteBytesContent(OutputPacket *pkt);
+
+  /**
+   * Used to invoke a write into the Socket, returns false if the socket is not
+   * ready for write
+   */
+  WriteState FlushWriteBuffer();
+
+  /**
+   * Set the socket to non-blocking mode
+   */
+  inline void SetNonBlocking(evutil_socket_t fd) {
+    auto flags = fcntl(fd, F_GETFL);
+    flags |= O_NONBLOCK;
+    if (fcntl(fd, F_SETFL, flags) < 0) {
+      LOG_ERROR("Failed to set non-blocking socket");
+    }
+  }
+
+  /**
+   * Set TCP No Delay for lower latency
+   */
+  inline void SetTCPNoDelay(evutil_socket_t fd) {
+    int one = 1;
+    setsockopt(fd, IPPROTO_TCP, TCP_NODELAY, &one, sizeof one);
+  }
+
+  int sock_fd_;                    // socket file descriptor
+  struct event *network_event = nullptr;  // something to read from network
+  struct event *workpool_event = nullptr; // worker thread done the job
+
+  SSL* conn_SSL_context = nullptr;          // SSL context for the connection
+
+  ConnectionHandlerTask *handler_;          // reference to the network thread
+  std::unique_ptr<ProtocolHandler> protocol_handler_;       // Stores state for this socket
+  tcop::TrafficCop traffic_cop_;
+
+  Buffer rbuf_;                     // Socket's read buffer
+  Buffer wbuf_;                     // Socket's write buffer
+  unsigned int next_response_ = 0;  // The next response in the response buffer
+  Client client_;
+  bool ssl_sent_ = false;
+  StateMachine state_machine_;
 };
+
 
 }
 }
