--- src/include/network/network_connection_factory.h
@@ -1,25 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-//                         Peloton
-//
-// network_connection_factory.h
-//
-// Identification: src/include/network/network_connection_factory.h
-//
-// Copyright (c) 2015-17, Carnegie Mellon University Database Group
-//
-//===----------------------------------------------------------------------===//
-
-#pragma once
-
-#include "network/network_connection.h"
-
-namespace peloton {
-namespace network {
-
-class NetworkConnectionFactory {
-
-};
-  
-}
-}
