+++ src/include/network/network_state.h
@@ -16,21 +16,15 @@ namespace peloton {
 namespace network {
 // Network Thread States
 enum class ConnState {
-  CONN_LISTENING,  // State that listens for new connections
-  CONN_READ,       // State that reads data from the network
-  CONN_WRITE,      // State the writes data to the network
-  CONN_WAIT,       // State for waiting for some event to happen
-  CONN_PROCESS,    // State that runs the network protocol on received data
-  CONN_CLOSING,    // State for closing the client connection
-  CONN_GET_RESULT, // State when triggered by worker thread that completes the task.
+  READ,       // State that reads data from the network
+  WRITE,      // State the writes data to the network
+  WAIT,       // State for waiting for some event to happen
+  PROCESS,    // State that runs the network protocol on received data
+  CLOSING,    // State for closing the client connection
+  GET_RESULT, // State when triggered by worker thread that completes the task.
 };
 
-enum class ReadState {
-  READ_DATA_RECEIVED,
-  READ_NO_DATA_RECEIVED,
-  READ_ERROR,
-};
-
+// TODO(tianyu): Convert use cases of this to just return Transition
 enum class WriteState {
   WRITE_COMPLETE,   // Write completed
   WRITE_NOT_READY,  // Socket not ready to write
